+++ packages/@tailwindcss-upgrade/src/template/is-safe-migration.ts
@@ -10,6 +10,7 @@ const CONDITIONAL_TEMPLATE_SYNTAX = [
   /x-if=['"]$/,
   /x-show=['"]$/,
 ]
+const NEXT_PLACEHOLDER_PROP = /placeholder=\{?['"]$/
 
 export function isSafeMigration(location: { contents: string; start: number; end: number }) {
   let currentLineBeforeCandidate = ''
@@ -61,6 +62,11 @@ export function isSafeMigration(location: { contents: string; start: number; end
     if (rule.test(currentLineBeforeCandidate)) {
       return false
     }
+  }
+
+  // Heuristic: Disallow Next.js Image `placeholder` prop
+  if (NEXT_PLACEHOLDER_PROP.test(currentLineBeforeCandidate)) {
+    return false
   }
 
   return true