+++ packages/tailwindcss/src/feature-flags.ts
@@ -1,0 +1,3 @@
+export const enableRemoveUnusedThemeVariables = false
+
+export const enableUserValid = process.env.FEATURES_ENV !== 'stable'