+++ packages/tailwindcss/src/intellisense.test.ts
@@ -526,3 +526,40 @@ test('Custom functional @utility', async () => {
   expect(classNames).toContain('example-xs')
   expect(classMap.get('example-xs')?.modifiers).toEqual(['foo', 'bar'])
 })
+
+test('Theme keys with underscores are suggested with underscores', async () => {
+  let input = css`
+    @import 'tailwindcss/utilities';
+
+    @theme {
+      /* Disable the spacing scale */
+      --spacing: initial;
+
+      /* This will get suggeted with a dot because its surrounded by numbers */
+      --spacing-1_5: 1.5rem;
+
+      /* This will get suggeted with a dot  */
+      --spacing-2\.5: 1.5rem;
+
+      /* This will get suggeted with an underscore */
+      --spacing-logo_margin: 0.875rem;
+    }
+  `
+
+  let design = await __unstable__loadDesignSystem(input, {
+    loadStylesheet: async (_, base) => ({
+      base,
+      content: '@tailwind utilities;',
+    }),
+  })
+
+  let entries = design.getClassList().filter(([name]) => name.startsWith('p-'))
+
+  expect(entries).toContainEqual(['p-1.5', { modifiers: [] }])
+  expect(entries).toContainEqual(['p-2.5', { modifiers: [] }])
+  expect(entries).toContainEqual(['p-logo_margin', { modifiers: [] }])
+
+  expect(entries).not.toContainEqual(['p-1_5', { modifiers: [] }])
+  expect(entries).not.toContainEqual(['p-2_5', { modifiers: [] }])
+  expect(entries).not.toContainEqual(['p-logo.margin', { modifiers: [] }])
+})